fix(striker-ui): relocate server menu file
